+++ .gitignore
@@ -10,6 +10,7 @@ __pycache__/
 
 # Config files
 botconfig.py
+/gamemodes.py
 
 # Database files
 *.sqlite3
