+++ gamemodes.py.example
@@ -1,0 +1,16 @@
+# To add new game modes, rename this file to 'gamemodes.py' then add them in
+# Basic frame for the game modes in src/gamemodes.py
+
+# *** DO NOT TOUCH ANY OF THE CODE BETWEEN THE LINES 4 AND 16 ***
+
+import src.settings as var
+
+from src.gamemodes import (
+
+    GameMode,
+    game_mode,
+    reset_roles,
+
+)
+
+# Add custom game modes from this point
